+++ pkg/wharfyml/builtins_test.go
@@ -39,12 +39,3 @@ func TestVarFilePrettyPath(t *testing.T) {
 
 	assert.Equal(t, want, got)
 }
-
-func TestUseShorthandHomeDir(t *testing.T) {
-	home := "/home/root"
-	path := "/home/root/.wharf-vars.yml"
-	want := "~/.wharf-vars.yml"
-	got := useShorthandHomePrefix(path, home)
-
-	assert.Equal(t, want, got)
-}
